+++ tests/test_analyze.py
@@ -86,6 +86,16 @@ Older: 1.0.1, 1.0.2, 1.0.3, 1.0.4, 1.0.5, 1.0.6, 1.0.7a1, 1.0.7b2
 """
         self.assertEqual(expected_report, report.render(verbose=True))
 
+    def test_report_ignore_all(self):
+        env = EnvPackages.from_freeze_file(StringIO('non-lts-example==1.0.7'))
+        version_db = VersionDB(yaml_db=TEST_DB_NAME)
+        available = FakePackageVersionInfo.from_list('non-lts-example', NON_LTS_EXAMPLE_VERSIONS)
+        analyzer = Analyzer(env, available, version_db)
+        report = analyzer.analyze(
+            ignored_packages=['non-lts-example']
+        )
+        self.assertEqual('', report.render())
+
     def test_report_no_newer(self):
         env = EnvPackages.from_freeze_file(StringIO('non-lts-example==1.0.8'))
         version_db = VersionDB(yaml_db=TEST_DB_NAME)